+++ test_unstructured_ingest/src/discord.sh
@@ -16,31 +16,31 @@ CI=${CI:-"false"}
 # shellcheck disable=SC1091
 source "$SCRIPT_DIR"/cleanup.sh
 function cleanup() {
-	cleanup_dir "$OUTPUT_DIR"
-	cleanup_dir "$WORK_DIR"
-	if [ "$CI" == "true" ]; then
-		cleanup_dir "$DOWNLOAD_DIR"
-	fi
+  cleanup_dir "$OUTPUT_DIR"
+  cleanup_dir "$WORK_DIR"
+  if [ "$CI" == "true" ]; then
+    cleanup_dir "$DOWNLOAD_DIR"
+  fi
 }
 trap cleanup EXIT
 
 if [ -z "$DISCORD_TOKEN" ]; then
-	echo "Skipping Discord ingest test because the DISCORD_TOKEN env var is not set."
-	exit 8
+  echo "Skipping Discord ingest test because the DISCORD_TOKEN env var is not set."
+  exit 8
 fi
 
 RUN_SCRIPT=${RUN_SCRIPT:-./unstructured/ingest/main.py}
 PYTHONPATH=${PYTHONPATH:-.} "$RUN_SCRIPT" \
-	discord \
-	--num-processes "$max_processes" \
-	--metadata-exclude coordinates,file_directory,metadata.data_source.date_processed,metadata.last_modified,metadata.detection_class_prob,metadata.parent_id,metadata.category_depth \
-	--download-dir "$DOWNLOAD_DIR" \
-	--preserve-downloads \
-	--reprocess \
-	--output-dir "$OUTPUT_DIR" \
-	--verbose \
-	--channels 1099442333440802930,1099601456321003600 \
-	--token "$DISCORD_TOKEN" \
-	--work-dir "$WORK_DIR"
+  discord \
+  --num-processes "$max_processes" \
+  --metadata-exclude coordinates,file_directory,metadata.data_source.date_processed,metadata.last_modified,metadata.detection_class_prob,metadata.parent_id,metadata.category_depth \
+  --download-dir "$DOWNLOAD_DIR" \
+  --preserve-downloads \
+  --reprocess \
+  --output-dir "$OUTPUT_DIR" \
+  --verbose \
+  --channels 1099442333440802930,1099601456321003600 \
+  --token "$DISCORD_TOKEN" \
+  --work-dir "$WORK_DIR"
 
 "$SCRIPT_DIR"/check-diff-expected-output.sh $OUTPUT_FOLDER_NAME