+++ test_unstructured_ingest/src/local.sh
@@ -14,23 +14,23 @@ max_processes=${MAX_PROCESSES:=$(python3 -c "import os; print(os.cpu_count())")}
 # shellcheck disable=SC1091
 source "$SCRIPT_DIR"/cleanup.sh
 function cleanup() {
-	cleanup_dir "$OUTPUT_DIR"
-	cleanup_dir "$WORK_DIR"
+  cleanup_dir "$OUTPUT_DIR"
+  cleanup_dir "$WORK_DIR"
 }
 trap cleanup EXIT
 
 RUN_SCRIPT=${RUN_SCRIPT:-./unstructured/ingest/main.py}
 PYTHONPATH=${PYTHONPATH:-.} "$RUN_SCRIPT" \
-	local \
-	--num-processes "$max_processes" \
-	--metadata-exclude coordinates,filename,file_directory,metadata.data_source.date_processed,metadata.last_modified,metadata.detection_class_prob,metadata.parent_id,metadata.category_depth \
-	--strategy hi_res \
-	--reprocess \
-	--output-dir "$OUTPUT_DIR" \
-	--verbose \
-	--file-glob "*.html" \
-	--input-path example-docs \
-	--work-dir "$WORK_DIR"
+  local \
+  --num-processes "$max_processes" \
+  --metadata-exclude coordinates,filename,file_directory,metadata.data_source.date_processed,metadata.last_modified,metadata.detection_class_prob,metadata.parent_id,metadata.category_depth \
+  --strategy hi_res \
+  --reprocess \
+  --output-dir "$OUTPUT_DIR" \
+  --verbose \
+  --file-glob "*.html" \
+  --input-path example-docs \
+  --work-dir "$WORK_DIR"
 
 "$SCRIPT_DIR"/check-num-files-output.sh 12 $OUTPUT_FOLDER_NAME